Mon, May 13, 2024, 7:36 PM +05:30

+++ readability/readability.c
@@ -28,7 +28,7 @@ int main(void)
     }
     else
     {
-        printf("Grade %i",round(index));
+        printf("Grade %i",(int)round(index));
     }
 
     printf("\n");
